Un-nest ComparerHelper<T> and DescendingComparer<T>
